+++ .env
@@ -1,3 +1,5 @@
 MONGODB_URI=mongodb://localhost:27017/banco
-PORT=3000
+AUTH_PORT=3001
+ACCOUNT_PORT=3002
+TRANSACTION_PORT=3003
 JWT_SECRET=hK92V!m#pG7$Lz!dT4vQ@eRf%Yx9JcWuZp&nHt$g
