+++ public/js/auth.js
@@ -96,7 +96,7 @@ async function handleLogin(e) {
     }
 
     try {
-        const response = await fetch('http://localhost:3000/api/login', {
+        const response = await fetch('http://localhost:3001/api/auth/login', {
             method: 'POST',
             headers: { 'Content-Type': 'application/json' },
             body: JSON.stringify({ email, contrasena: password })
